@@ -3,23 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@tsparticles/react": "^3.0.0",
     "axios": "^0.21.2",
     "classnames": "^2.2.5",
     "connected-react-router": "^6.5.2",
     "d3": "^5.1.0",
     "history": "^4.10.1",
     "lodash": "^4.17.21",
-    "particles.js": "^2.0.0",
     "prop-types": "^15.6.0",
     "query-string": "^7.0.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-autosuggest": "^10.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.0.0",
     "react-lazyload": "^3.0.0",
     "react-loadable": "^5.3.1",
     "react-modal": "^3.3.1",
-    "react-particles-js": "^3.2.1",
     "react-redux": "^7.0.3",
     "react-router-dom": "^5.1.2",
     "react-scripts": "5.0.1",
@@ -27,6 +26,7 @@
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
     "styled-components": "^5.0.1",
+    "tsparticles": "^3.0.2",
     "typeface-crimson-text": "1.1.13",
     "typeface-quicksand": "1.1.13"
   },
